Typo, kommentezés

+++ app/src/main/java/hu/bme/aut/szoftarch/kozkincsker/data/model/Feedback.kt
@@ -4,13 +4,21 @@ import android.os.Parcelable
 import com.google.firebase.firestore.PropertyName
 import kotlinx.parcelize.Parcelize
 import java.util.UUID
-
+/**
+ * Küldetés visszajelzés osztály
+ * @param id
+ * @param mission kapcsolódó küldetés
+ * @param writer értékelést író felhasználó
+ * @param stars 0-5 skálán csillaggal való küldetés értékelés
+ * @param comment szöveges küldetés értékelés
+ * @param commentState szöveges értékelés láthatósága
+ */
 @Parcelize
 data class Feedback(
     var id: String = UUID.randomUUID().toString(),
     var mission: Mission = Mission(),
     var writer: User = User(),
-    var stars: Double = 0.0,
+    var stars: Int = 0,
     var comment: String = "",
     var commentState: CommentVisibility = CommentVisibility.SHOW
 ): Parcelable{
